Refactor apy compute types
